Move install logic to it's original place.
