+++ package.json
@@ -42,6 +42,7 @@
     "request": "2.74.0",
     "unzipper": "^0.1.11",
     "ansi_up": "1.3.0",
+    "colors": "1.1.2",
     "semver": "5.3.0",
     "async": "1.5.2",
     "gulp": "3.9.1",
